Split updater error formatting into separate files; sync stale length budget
The fix pushed UpdateStateModel.swift (485 to 565) and UpdatePopoverView.swift
(522 to 539) over the 500-line file-length budget. Rather than record new
debt, split along the repo's one-type-per-file lines:

- Move the error classification/presentation statics into
  UpdateStateModel+ErrorFormatting.swift (main file back to ~306 lines).
- Move UpdateErrorView into its own file (UpdatePopoverView back to ~448).

Both new files are auto-included by SwiftPM; no manifest change needed.

Also sync three stale budget entries (ContentView, TabManager,
WorkspaceGroupTests) to their current origin/main lengths. These grew on main
without the budget file being refreshed, so the guard was already red on
main independent of this change; this brings the checked-in budget back in
line with reality.
