+++ Packages/CmuxUpdater/Sources/CmuxUpdater/UpdateStateModel.swift
@@ -298,265 +298,6 @@ public final class UpdateStateModel {
         return String(localized: "update.available.withVersion", defaultValue: "Update Available: \(version)")
     }
 
-    // MARK: - Error formatting
-
-    /// A short, user-facing title for an update error.
-    public static func userFacingErrorTitle(for error: any Swift.Error) -> String {
-        let nsError = error as NSError
-        if let networkError = networkError(from: nsError) {
-            switch networkError.code {
-            case NSURLErrorNotConnectedToInternet:
-                return String(localized: "update.error.noInternet.title", defaultValue: "No Internet Connection")
-            case NSURLErrorTimedOut:
-                return String(localized: "update.error.timedOut.title", defaultValue: "Update Timed Out")
-            case NSURLErrorCannotFindHost:
-                return String(localized: "update.error.serverNotFound.title", defaultValue: "Server Not Found")
-            case NSURLErrorCannotConnectToHost:
-                return String(localized: "update.error.serverUnreachable.title", defaultValue: "Server Unreachable")
-            case NSURLErrorNetworkConnectionLost:
-                return String(localized: "update.error.connectionLost.title", defaultValue: "Connection Lost")
-            case NSURLErrorSecureConnectionFailed,
-                 NSURLErrorServerCertificateUntrusted,
-                 NSURLErrorServerCertificateHasBadDate,
-                 NSURLErrorServerCertificateHasUnknownRoot,
-                 NSURLErrorServerCertificateNotYetValid:
-                return String(localized: "update.error.secureConnectionFailed.title", defaultValue: "Secure Connection Failed")
-            default:
-                break
-            }
-        }
-        if isUpdaterAgentConnectionFailure(nsError) {
-            return String(localized: "update.error.installerAgent.title", defaultValue: "Couldn’t Start Updater")
-        }
-        if nsError.domain == SUSparkleErrorDomain {
-            switch nsError.code {
-            case 4005:
-                return String(localized: "update.error.installFailed.title", defaultValue: "Couldn’t Install Update")
-            case 2001:
-                return String(localized: "update.error.downloadFailed.title", defaultValue: "Couldn't Download Update")
-            case 1000, 1002:
-                return String(localized: "update.error.feedError.title", defaultValue: "Update Feed Error")
-            case 4:
-                return String(localized: "update.error.invalidFeed.title", defaultValue: "Invalid Update Feed")
-            case 3:
-                return String(localized: "update.error.insecureFeed.title", defaultValue: "Insecure Update Feed")
-            case 1, 2, 3001, 3002:
-                return String(localized: "update.error.signatureError.title", defaultValue: "Update Signature Error")
-            case 1003, 1005:
-                return String(localized: "update.error.appLocation.title", defaultValue: "App Location Issue")
-            default:
-                break
-            }
-        }
-        return String(localized: "update.error.failed.title", defaultValue: "Update Failed")
-    }
-
-    /// A user-facing explanatory message for an update error.
-    public static func userFacingErrorMessage(for error: any Swift.Error) -> String {
-        let nsError = error as NSError
-        if let networkError = networkError(from: nsError) {
-            switch networkError.code {
-            case NSURLErrorNotConnectedToInternet:
-                return String(localized: "update.error.noInternet.message", defaultValue: "cmux can’t reach the update server. Check your internet connection and try again.")
-            case NSURLErrorTimedOut:
-                return String(localized: "update.error.timedOut.message", defaultValue: "The update server took too long to respond. Try again in a moment.")
-            case NSURLErrorCannotFindHost:
-                return String(localized: "update.error.serverNotFound.message", defaultValue: "The update server can’t be found. Check your connection or try again later.")
-            case NSURLErrorCannotConnectToHost:
-                return String(localized: "update.error.serverUnreachable.message", defaultValue: "cmux couldn’t connect to the update server. Check your connection or try again later.")
-            case NSURLErrorNetworkConnectionLost:
-                return String(localized: "update.error.connectionLost.message", defaultValue: "The network connection was lost while checking for updates. Try again.")
-            case NSURLErrorSecureConnectionFailed,
-                 NSURLErrorServerCertificateUntrusted,
-                 NSURLErrorServerCertificateHasBadDate,
-                 NSURLErrorServerCertificateHasUnknownRoot,
-                 NSURLErrorServerCertificateNotYetValid:
-                return String(localized: "update.error.secureConnectionFailed.message", defaultValue: "A secure connection to the update server couldn’t be established. Try again later.")
-            default:
-                break
-            }
-        }
-        if isUpdaterAgentConnectionFailure(nsError) {
-            return String(localized: "update.error.installerAgent.message", defaultValue: "macOS couldn’t start the updater service, which can happen after your Mac has been running for a long time. Restart your Mac and try again, or download the latest version below.")
-        }
-        if nsError.domain == SUSparkleErrorDomain {
-            switch nsError.code {
-            case 2001:
-                return String(localized: "update.error.feedDownload.message", defaultValue: "cmux couldn't download the update feed. Check your connection and try again.")
-            case 1000, 1002:
-                return String(localized: "update.error.feedRead.message", defaultValue: "The update feed could not be read. Please try again later.")
-            case 4:
-                return String(localized: "update.error.invalidFeed.message", defaultValue: "The update feed URL is invalid. Please contact support.")
-            case 3:
-                return String(localized: "update.error.insecureFeed.message", defaultValue: "The update feed is insecure. Please contact support.")
-            case 1, 2, 3001, 3002:
-                return String(localized: "update.error.signatureError.message", defaultValue: "The update's signature could not be verified. Please try again later.")
-            case 1003, 1005:
-                return String(localized: "update.error.permissionError.message", defaultValue: "Move cmux into Applications and relaunch to enable updates.")
-            case 4005:
-                return String(localized: "update.error.installFailed.message", defaultValue: "The update couldn’t be installed. Restart your Mac and try again, or download the latest version below.")
-            default:
-                break
-            }
-        }
-        // Catch-all: keep user-facing copy in cmux terms; raw vendor descriptions, domains, and
-        // codes stay in `errorDetails` (the copyable Details block + the update log), not here.
-        return String(localized: "update.error.failed.message", defaultValue: "Something went wrong while checking for updates. Try again, or check the update log for details.")
-    }
-
-    /// A direct download URL for the latest release when manually downloading is a sensible
-    /// recovery for `error`, or `nil` when it is not.
-    ///
-    /// Returned for installation, extraction, resume, and download failures, where grabbing the
-    /// latest build sidesteps a broken in-app install (for example the wedged-launchd case this
-    /// surfaces a "Download Latest Version" button for). Returns `nil` for feed, signature,
-    /// configuration, and "already up to date" errors, where a manual download would not help or
-    /// could be unsafe (notably signature/validation failures).
-    public static func manualDownloadURL(for error: any Swift.Error) -> URL? {
-        let nsError = error as NSError
-        guard nsError.domain == SUSparkleErrorDomain else { return nil }
-        switch nsError.code {
-        case 1004,                                    // SUResumeAppcastError
-             2000, 2001,                              // temp-directory / download failures
-             3000,                                    // SUUnarchivingError
-             4000, 4001, 4002, 4003, 4004, 4005, 4006, // file copy / auth / installer failures
-             4010, 4012:                              // agent invalidation / write-permission failures
-            return URL(string: manualDownloadURLString)
-        default:
-            return nil
-        }
-    }
-
-    /// Builds the multi-line technical detail block shown in the error popover.
-    ///
-    /// - Parameters:
-    ///   - error: The error to describe.
-    ///   - technicalDetails: Extra detail captured at failure time, if any.
-    ///   - feedURLString: The feed URL in effect at failure time, if any.
-    ///   - logPath: The path of the update log file (from ``UpdateLogging/logPath()``), appended
-    ///     so users can find the full trace.
-    /// - Returns: A newline-separated detail block.
-    public static func errorDetails(for error: any Swift.Error,
-                                    technicalDetails: String?,
-                                    feedURLString: String?,
-                                    logPath: String) -> String {
-        let nsError = error as NSError
-        var lines: [String] = []
-        lines.append("Message: \(nsError.localizedDescription)")
-        lines.append("Domain: \(nsError.domain)")
-        if nsError.domain == SUSparkleErrorDomain,
-           let sparkleName = sparkleErrorCodeName(for: nsError.code) {
-            lines.append("Code: \(sparkleName) (\(nsError.code))")
-        } else {
-            lines.append("Code: \(nsError.code)")
-        }
-
-        if let url = nsError.userInfo[NSURLErrorFailingURLErrorKey] as? URL {
-            lines.append("URL: \(url.absoluteString)")
-        } else if let urlString = nsError.userInfo[NSURLErrorFailingURLStringErrorKey] as? String {
-            lines.append("URL: \(urlString)")
-        }
-
-        if let failure = nsError.userInfo[NSLocalizedFailureReasonErrorKey] as? String,
-           !failure.isEmpty {
-            lines.append("Failure: \(failure)")
-        }
-        if let recovery = nsError.userInfo[NSLocalizedRecoverySuggestionErrorKey] as? String,
-           !recovery.isEmpty {
-            lines.append("Recovery: \(recovery)")
-        }
-
-        if let feedURLString, !feedURLString.isEmpty {
-            lines.append("Feed: \(feedURLString)")
-        }
-
-        if let technicalDetails, !technicalDetails.isEmpty {
-            lines.append("Debug: \(technicalDetails)")
-        }
-
-        lines.append("Log: \(logPath)")
-        return lines.joined(separator: "\n")
-    }
-
-    /// The canonical direct-download URL for the latest stable release artifact, used as the
-    /// manual fallback when Sparkle's in-app install path fails.
-    private static let manualDownloadURLString = "https://github.com/manaflow-ai/cmux/releases/latest/download/cmux-macos.dmg"
-
-    /// Whether an error reflects Sparkle's updater helper agent never connecting.
-    ///
-    /// The login session's launchd domain can wedge into on-demand-only mode after a very long
-    /// uptime, so launchd refuses to spawn Sparkle's installer/progress agent and the install
-    /// times out ("agent connection was never initiated"). This surfaces as ``SUInstallationError``
-    /// (4005) wrapping an underlying Sparkle error, or as ``SUAgentInvalidationError`` (4010).
-    /// Restarting the Mac recreates the session domain and clears the wedge, so the user-facing
-    /// copy points there rather than at the misleading "move into Applications" guidance.
-    private static func isUpdaterAgentConnectionFailure(_ error: NSError) -> Bool {
-        guard error.domain == SUSparkleErrorDomain else { return false }
-        if error.code == 4010 { return true }
-        guard error.code == 4005 else { return false }
-        if let underlying = error.userInfo[NSUnderlyingErrorKey] as? NSError,
-           underlying.domain == SUSparkleErrorDomain {
-            return true
-        }
-        let text = [
-            error.localizedDescription,
-            (error.userInfo[NSLocalizedFailureReasonErrorKey] as? String) ?? "",
-            (error.userInfo[NSLocalizedRecoverySuggestionErrorKey] as? String) ?? "",
-        ].joined(separator: "\n").lowercased()
-        return text.contains("agent connection") || text.contains("remote port")
-    }
-
-    private static func networkError(from error: NSError) -> NSError? {
-        if error.domain == NSURLErrorDomain {
-            return error
-        }
-        if let underlying = error.userInfo[NSUnderlyingErrorKey] as? NSError,
-           underlying.domain == NSURLErrorDomain {
-            return underlying
-        }
-        return nil
-    }
-
-    private static func sparkleErrorCodeName(for code: Int) -> String? {
-        switch code {
-        case 1: return "SUNoPublicDSAFoundError"
-        case 2: return "SUInsufficientSigningError"
-        case 3: return "SUInsecureFeedURLError"
-        case 4: return "SUInvalidFeedURLError"
-        case 5: return "SUInvalidUpdaterError"
-        case 6: return "SUInvalidHostBundleIdentifierError"
-        case 7: return "SUInvalidHostVersionError"
-        case 1000: return "SUAppcastParseError"
-        case 1001: return "SUNoUpdateError"
-        case 1002: return "SUAppcastError"
-        case 1003: return "SURunningFromDiskImageError"
-        case 1004: return "SUResumeAppcastError"
-        case 1005: return "SURunningTranslocated"
-        case 1006: return "SUWebKitTerminationError"
-        case 1007: return "SUReleaseNotesError"
-        case 2000: return "SUTemporaryDirectoryError"
-        case 2001: return "SUDownloadError"
-        case 3000: return "SUUnarchivingError"
-        case 3001: return "SUSignatureError"
-        case 3002: return "SUValidationError"
-        case 4000: return "SUFileCopyFailure"
-        case 4001: return "SUAuthenticationFailure"
-        case 4002: return "SUMissingUpdateError"
-        case 4003: return "SUMissingInstallerToolError"
-        case 4004: return "SURelaunchError"
-        case 4005: return "SUInstallationError"
-        case 4006: return "SUDowngradeError"
-        case 4007: return "SUInstallationCanceledError"
-        case 4008: return "SUInstallationAuthorizeLaterError"
-        case 4009: return "SUNotValidUpdateError"
-        case 4010: return "SUAgentInvalidationError"
-        case 4012: return "SUInstallationWriteNoPermissionError"
-        case 5000: return "SUIncorrectAPIUsageError"
-        default:
-            return nil
-        }
-    }
-
     /// Normalizes a Sparkle display version into a trimmed, non-empty string, or `nil`.
     public static func normalizedDetectedUpdateVersion(from version: String) -> String? {
         let trimmed = version.trimmingCharacters(in: .whitespacesAndNewlines)
